README
The importance of Screw Theory in robotics is recognized, but in practice, very few teach it to engineering students and therefore, not many postgraduates know how to use it. However, in a variety of areas of robotics, the methods and formalisms based on the geometry and algebra of the screws have proven to be superior to other techniques and have led to significant advances.